+++ freight/__init__.py
@@ -1,4 +1,4 @@
 default_app_config = 'freight.apps.FreightConfig'
 
-__version__ = '0.12.5'
+__version__ = '0.13.0'
 __title__ = 'Alliance Freight'